pyproject.toml
[project]
name = "caloru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-filter>=25.2",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "ruff>=0.13.1",
]
prod = [
    "psycopg2",
]
